+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 2025-01-08
+### Added
+- Variable `$ITEMNAME`
+- Logs isLoading state
 
 ## [0.1.0] - 2025-01-03
 ### Added